requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.4
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.7
Pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
django-cors-headers==4.3.1
